backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.35.0
pydantic==2.10.6
pydantic-settings==2.8.1
psycopg[binary]==3.2.5
psycopg-pool==3.2.6
httpx==0.28.1
python-jose[cryptography]==3.3.0
boto3==1.35.90
